requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
iniconfig==2.0.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pendulum==3.0.0
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
time-machine==2.16.0
tzdata==2024.2
urllib3==2.2.3
